requirements.txt
Django==3.2.12
psycopg2==2.9.5
Pillow==10.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.2.0
websockets==12.0